refactoring swtc-serializer with typescript
