+++ packages/swtc-serializer/.travis.yml
@@ -4,4 +4,4 @@ node_js:
 script:
   - npm run test
   - npm run prettier:check
-  - npm run eslint
+  - npm run tslint
