+++ packages/swtc-serializer/src/DataCheck.ts
@@ -16,82 +16,70 @@
  *
  * Data functions used to check the valid data types.
  */
-var CURRENCY_NAME_LEN = 3
-var CURRENCY_NAME_LEN2 = 6
-var TUM_NAME_LEN = 40
-var WalletFactory = require("swtc-wallet").Factory
 
-function Factory(Wallet = WalletFactory()) {
+import { Factory as WalletFactory } from "swtc-wallet"
+import { MAX_CURRENCY_LEN, MIN_CURRENCY_LEN, TUM_NAME_LEN } from "./Constant"
+
+const Factory = (Wallet = WalletFactory("jingtum")) => {
   if (!Wallet.hasOwnProperty("KeyPair")) {
     throw Error("Datacheck need a Wallet class")
   }
 
-  function allNumeric(text) {
+  const allNumeric = (text: any): boolean => {
     // assign a string with numbers (0-9) in the HTML form
-    var numbers = /^[0-9]+$/
+    const numbers = /^[0-9]+$/
     // Check if the input contains all numbers.
-    if (String(text).match(numbers)) {
-      return true
-    } else {
-      return false
-    }
+    return Boolean(String(text).match(numbers))
   }
 
   /*
    * Decide if the input is a valid string
    * for a float value.
    */
-  function isFloat(val) {
-    var floatRegex = /^-?\d+(?:[.,]\d*?)?$/
-    if (!floatRegex.test(val) || Number.isNaN(parseFloat(val))) {
-      return false
-    }
-
-    return true
+  const isFloat = (val: any): boolean => {
+    const floatRegex = /^-?\d+(?:[.,]\d*?)?$/
+    return Boolean(floatRegex.test(val) && !Number.isNaN(parseFloat(val)))
   }
 
   // Detect if the string contains only
   // numbers and capital letters.
-
-  function isLetterNumer(str) {
-    var numbers = /^[0-9A-Z]+$/
-    return !!String(str).match(numbers)
+  const isLetterNumer = (str: any): boolean => {
+    const numbers = /^[0-9A-Z]+$/i
+    return Boolean(String(str).match(numbers))
   }
 
   // return true if the code is 3 letters/numbers
-  function isCurrency(code) {
+  const isCurrency = (code: any): boolean => {
     return (
       typeof code === "string" &&
-      (!!code &&
-        code.length >= CURRENCY_NAME_LEN &&
-        code.length <= CURRENCY_NAME_LEN2)
+      Boolean(
+        code &&
+          code.length >= MIN_CURRENCY_LEN &&
+          code.length <= MAX_CURRENCY_LEN
+      )
     )
   }
 
   // return true if the code is 40 letters/numbers
-  function isCustomTum(code) {
+  const isCustomTum = (code: any): boolean => {
     return isLetterNumer(code) && String(code).length === TUM_NAME_LEN
   }
 
   /*
    * Return true is the input string
    * is a valid TUM code
-   *
    */
-  function isTumCode(code) {
+  const isTumCode = (code: any): boolean => {
     // input must be defined and non-null
     // Make sure if the code meets the coding rule
     // tum: Custom tum, 40 capital letters or number
-    return (
-      typeof code === "string" &&
-      (code === "SWT" || isCurrency(code) || isCustomTum(code))
-    )
+    return typeof code === "string" && (isCurrency(code) || isCustomTum(code))
   }
 
   /*
    * Only valid for freeze and authorize
    */
-  function isRelation(str) {
+  const isRelation = (str: any): boolean => {
     return typeof str === "string" && (str === "freeze" || str === "autorize")
   }
 
@@ -102,31 +90,30 @@ function Factory(Wallet = WalletFactory()) {
    * issuer/counterparty
    * currency
    */
-  function isAmount(obj) {
-    if (obj === null || typeof obj !== "object") {
-      return false
-    }
-    if (typeof obj.value !== "string" || !isFloat(obj.value)) {
-      return false
-    }
-    if (!isTumCode(obj.currency)) {
+  const isAmount = (obj: any): boolean => {
+    if (
+      obj === null ||
+      typeof obj !== "object" ||
+      typeof obj.value !== "string" ||
+      !isFloat(obj.value) ||
+      !isTumCode(obj.currency)
+    ) {
       return false
     }
     // AMOUNT could have a field named
     // either as 'issuer'
     // or as 'counterparty'
     // for SWT, this can be undefined
-    if (typeof obj.issuer !== "undefined" && obj.issuer !== "") {
+    if (obj.issuer) {
       if (!Wallet.isValidAddress(obj.issuer)) {
         return false
       }
     } else {
-      // if currency === 'SWT',自动补全issuer.
-      if (obj.currency === "SWT") {
-        obj.issuer = ""
-      } else {
+      if (obj.currency !== "SWT") {
         return false
       }
+      // if currency === 'SWT',自动补全issuer.
+      obj.issuer = ""
     }
     return true
   }
@@ -138,27 +125,19 @@ function Factory(Wallet = WalletFactory()) {
    * counterparty  String  货币发行方
    * freezed  String  冻结的金额
    */
-  function isBalance(obj) {
-    if (obj === null || typeof obj !== "object") {
+  const isBalance = (obj: any): boolean => {
+    if (
+      obj === null ||
+      typeof obj !== "object" ||
+      !isFloat(obj.freezed) ||
+      !isFloat(obj.value) ||
+      !isTumCode(obj.currency)
+    ) {
       return false
     }
-
-    if (!isFloat(obj.freezed)) {
-      return false
-    }
-
-    if (!isFloat(obj.value)) {
-      return false
-    }
-
-    if (!isTumCode(obj.currency)) {
-      return false
-    }
-
     // AMOUNT could have a field named
     // either as 'issuer'
     // or as 'counterparty'
-
     if (!Wallet.isValidAddress(obj.counterparty)) {
       return false
     }
@@ -166,8 +145,8 @@ function Factory(Wallet = WalletFactory()) {
   }
 
   return {
-    isCustomTum,
     allNumeric,
+    isCustomTum,
     isRelation,
     isTumCode,
     isCurrency,
@@ -176,4 +155,5 @@ function Factory(Wallet = WalletFactory()) {
     isLetterNumer
   }
 }
-module.exports = { Factory }
+
+export { Factory }
