+++ packages/swtc-serializer/src/model/index.ts
@@ -1,0 +1,5 @@
+export interface IAmount {
+  currency: string
+  issuer?: string
+  value: string
+}
