+++ packages/swtc-serializer/src/types/STHash160.ts
@@ -1,0 +1,12 @@
+import { convertHexToByteArray, serializeHex } from "../Utils"
+import SerializedType from "./SerializedType"
+const STHash160 = new SerializedType({
+  serialize(so, val) {
+    serializeHex(so, convertHexToByteArray(val), true)
+  },
+  parse(so) {
+    return this.KeyPair.convertBytesToAddress(so.read(20))
+  }
+})
+
+export default STHash160
