+++ packages/swtc-serializer/src/types/STInt16.ts
@@ -1,0 +1,12 @@
+import { convertIntegerToByteArray, readAndSum } from "../Utils"
+import SerializedType from "./SerializedType"
+const STInt16 = new SerializedType({
+  serialize(so, val) {
+    so.append(convertIntegerToByteArray(val, 2))
+  },
+  parse(so) {
+    return readAndSum(so, 2)
+  }
+})
+
+export default STInt16
