+++ packages/swtc-serializer/src/types/STVL.ts
@@ -1,0 +1,18 @@
+import { convertByteArrayToHex, isString, serializeHex } from "../Utils"
+import SerializedType from "./SerializedType"
+
+const STVL = new SerializedType({
+  serialize(so, val) {
+    if (isString(val)) {
+      serializeHex(so, val)
+    } else {
+      throw new Error("Unknown datatype.")
+    }
+  },
+  parse(so) {
+    const len = this.parse_varint(so)
+    return convertByteArrayToHex(so.read(len))
+  }
+})
+
+export default STVL
