--- packages/swtc-serializer/test/config.js
@@ -1,8 +1,0 @@
-module.exports = {
-    JT_NODE: 'ws://101.200.176.238:5020',
-    TEST_NODE: 'ws://101.200.176.238:5020',
-    testAddress: 'jfdqBEDsbk3eMSXX2t7CGeu2RPkEjHs6ie',
-    testSecret: 'shVCQFSxkF7DLXkrHY8X2PBKCKxS9',
-    testDestinationAddress: 'jsUVET1DkW5KuVeP5TR2GZZWQPFTqtxW9V',
-    testCreateHash: '20753B803666F729F99B3F2E90AD4E9731572D773B0E3E0DEB733197196F4EB4'
-}
